Fix bug #1583

+++ core/unit_test/TestView_64bit.hpp
@@ -99,7 +99,7 @@ void test_64bit(){
   }
   {
     int N = 1024*1024*1500;
-    int P = 1713091;
+    int64_t P = 1713091;
     Kokkos::View<int*, Device> a("A",N);
     Kokkos::parallel_for("FillA",Kokkos::RangePolicy<Kokkos::IndexType<int>>(0,N), KOKKOS_LAMBDA(const int& i) {
       a(i) = i%P;
@@ -108,7 +108,7 @@ void test_64bit(){
     Kokkos::parallel_reduce("FillA",Kokkos::RangePolicy<Kokkos::IndexType<int>>(0,N), KOKKOS_LAMBDA(const int& i,int64_t& lsum) {
       lsum += a(i);
     },sum);
-    int64_t expected = P*(P-1)/2 * (N/P) + (N%P)*(N%P-1)/2;
+    int64_t expected = (P*(P-1)/2) * int64_t(N/P) + (N%P)*(N%P-1)/2;
     ASSERT_EQ(expected,sum);
   }
 }
